Assistente_Viagens/requirements.txt
langchain_core
langchain_openai
langchain_community
python-dotenv